Test Rest Product Teamplate by Junit Test, LocalServerPort, WebEnvironment

+++ src/main/java/com/tma/api/repository/ProductRepository.java
@@ -3,6 +3,6 @@ package com.tma.api.repository;
 import com.tma.api.domain.Product;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends CrudRepository<Product, Long> {
+public interface ProductRepository extends CrudRepository<Product, Integer> {
 
 }
